+++ Tests/MapboxSearchTests/Common/Data Samples/CoreUserActivityReporterStub.swift
@@ -4,9 +4,13 @@ import Foundation
 @testable import MapboxSearch
 
 final class CoreUserActivityReporterStub: CoreUserActivityReporterProtocol {
+    var passedActivity: String?
+
     static func getOrCreate(for options: CoreUserActivityReporterOptions) -> CoreUserActivityReporter {
         preconditionFailure()
     }
     
-    func reportActivity(forComponent activity: String) {}
+    func reportActivity(forComponent activity: String) {
+        passedActivity = activity
+    }
 }